directory structure 수정
