+++ src/main/java/com/potato/balbambalbam/MyCommandLineRunner.java
@@ -2,7 +2,6 @@ package com.potato.balbambalbam;
 
 import com.potato.balbambalbam.main.cardList.service.UpdatePhonemeService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
