+++ src/main/java/com/potato/balbambalbam/data/entity/CardWeakSoundId.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.entity;
+package com.potato.balbambalbam.data.entity;
 
 import java.io.Serializable;
