+++ src/main/java/com/potato/balbambalbam/data/entity/Category.java
@@ -1,4 +1,4 @@
-package com.potato.balbambalbam.entity;
+package com.potato.balbambalbam.data.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
