+++ src/main/java/com/potato/balbambalbam/data/repository/CardRepository.java
@@ -1,6 +1,6 @@
-package com.potato.balbambalbam.main.repository;
+package com.potato.balbambalbam.data.repository;
 
-import com.potato.balbambalbam.entity.Card;
+import com.potato.balbambalbam.data.entity.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
