+++ src/main/java/com/potato/balbambalbam/data/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package com.potato.balbambalbam.main.repository;
+package com.potato.balbambalbam.data.repository;
 
-import com.potato.balbambalbam.entity.Category;
+import com.potato.balbambalbam.data.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
