+++ src/main/java/com/potato/balbambalbam/data/repository/UserWeakSoundRepository.java
@@ -1,6 +1,6 @@
-package com.potato.balbambalbam.main.repository;
+package com.potato.balbambalbam.data.repository;
 
-import com.potato.balbambalbam.entity.UserWeakSound;
+import com.potato.balbambalbam.data.entity.UserWeakSound;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
