+++ src/main/java/com/potato/balbambalbam/main/cardFeedback/dto/AiFeedbackResponseDto.java
@@ -1,6 +1,5 @@
 package com.potato.balbambalbam.main.cardFeedback.dto;
 
-import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
