+++ src/main/java/com/potato/balbambalbam/main/cardFeedback/dto/UserFeedbackRequestDto.java
@@ -1,10 +1,8 @@
 package com.potato.balbambalbam.main.cardFeedback.dto;
 
 import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 
 @Getter
 @NoArgsConstructor
