+++ src/main/java/com/potato/balbambalbam/main/cardList/service/CardListService.java
@@ -1,20 +1,19 @@
 package com.potato.balbambalbam.main.cardList.service;
 
-import com.potato.balbambalbam.entity.Card;
-import com.potato.balbambalbam.entity.CardBookmark;
-import com.potato.balbambalbam.entity.CardScore;
-import com.potato.balbambalbam.entity.Category;
+import com.potato.balbambalbam.data.entity.Card;
+import com.potato.balbambalbam.data.entity.CardBookmark;
+import com.potato.balbambalbam.data.entity.CardScore;
+import com.potato.balbambalbam.data.entity.Category;
+import com.potato.balbambalbam.data.repository.*;
 import com.potato.balbambalbam.main.cardList.dto.ResponseCardDto;
 import com.potato.balbambalbam.main.cardList.exception.CardNotFoundException;
 import com.potato.balbambalbam.main.cardList.exception.CategoryNotFoundException;
-import com.potato.balbambalbam.main.repository.*;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Supplier;
 
 @Service
 @RequiredArgsConstructor
